Align OVAL to styleguide and rename objects
The intention is the make it more readable. However, some improvements
were already done while reviewing. The regex to detect log files was
improved by escaping one literal "/". It was also removed a filter which
was not necessary. After this, the rules are no longer reporting
"notchecked" and we can use the test scenarios to test it.